requirements.txt
streamlit
google-cloud-texttospeech
google-auth

